+++ NEWS.md
@@ -1,5 +1,10 @@
 # MsFeatures 0.0
 
+## MsFeatures 0.0.4
+
+- Add `AbundanceSimilarityParam` and functionality to group features based on
+  similarity of abundances across samples.
+
 ## MsFeatures 0.0.3
 
 - Add functionality to group features based on similar retention times.